+++ BoatRentalSystem.Core/Entities/Owner.cs
@@ -15,7 +15,7 @@ namespace BoatRentalSystem.Core.Entities
 
         public string? Address { get; set; }
 
-        public decimal WalletBalance { get; set; } = 0.0m;
+        public double WalletBalance { get; set; } = 0.0;
 
         public DateTime CreatedAt { get; set; }= DateTime.Now;
